+++ src/main/kotlin/com/back/domain/review/dto/ReviewWriteRequest.kt
@@ -9,5 +9,5 @@ data class ReviewWriteRequest(
     @field:NotBlank val title: String,
     @field:NotBlank val content: String,
     @field:NotNull @field:DecimalMin("0") @field:DecimalMax("5") val rating: Double,
-    @field:NotNull val gameId: Int,
+    @field:NotNull val gameId: Long,
 )
